refactor for database directory
