+++ backend/app/main.py
@@ -2,7 +2,7 @@
 from fastapi import FastAPI
 from backend.app.routes.analyze import router as analyze_router
 from backend.app.routes.detections import router as detections_router
-from backend.app.database import Base, engine
+from database.config import Base, engine
 from backend.app.models.detection import Detections
 
 # Initialize FastAPI app
